+++ src/main/java/roomescape/payment/error/exception/PaymentClientException.java
@@ -1,8 +1,8 @@
-package roomescape.reservation.error.exception;
+package roomescape.payment.error.exception;
 
 import lombok.Getter;
 import org.springframework.http.HttpStatusCode;
-import roomescape.reservation.dto.response.PaymentErrorResponse;
+import roomescape.payment.dto.PaymentErrorResponse;
 
 @Getter
 public class PaymentClientException extends RuntimeException {
