+++ src/main/java/roomescape/payment/error/handler/PaymentExceptionHandler.java
@@ -1,12 +1,12 @@
-package roomescape.reservation.error.handler;
+package roomescape.payment.error.handler;
 
 import org.springframework.core.annotation.Order;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
 import roomescape.reservation.dto.response.ErrorResponse;
-import roomescape.reservation.error.exception.PaymentClientException;
-import roomescape.reservation.error.exception.PaymentServerException;
+import roomescape.payment.error.exception.PaymentClientException;
+import roomescape.payment.error.exception.PaymentServerException;
 
 @RestControllerAdvice
 @Order(1)
