+++ src/main/java/roomescape/payment/repository/PaymentRepository.java
@@ -1,7 +1,7 @@
-package roomescape.reservation.repository;
+package roomescape.payment.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import roomescape.reservation.entity.Payment;
+import roomescape.payment.entity.Payment;
 
 public interface PaymentRepository extends JpaRepository<Payment, Long> {
 }
