+++ src/main/java/roomescape/payment/service/PaymentRestClient.java
@@ -1,13 +1,12 @@
-package roomescape.reservation.service;
+package roomescape.payment.service;
 
 import java.util.Base64;
 import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Component;
 import org.springframework.web.client.RestClient;
-import roomescape.reservation.dto.response.PaymentApproveResponse;
-import roomescape.reservation.entity.Payment;
-import roomescape.reservation.error.handler.PaymentResponseErrorHandler;
+import roomescape.payment.entity.Payment;
+import roomescape.payment.error.handler.PaymentResponseErrorHandler;
 
 @Component
 @RequiredArgsConstructor
@@ -26,7 +25,7 @@ public class PaymentRestClient {
                 .body(payment)
                 .retrieve()
                 .onStatus(paymentResponseErrorHandler)
-                .body(PaymentApproveResponse.class);
+                .body(Payment.class);
     }
 
     private String getEncodedKey() {
