--- src/main/java/roomescape/reservation/service/PaymentService.java
@@ -1,21 +1,0 @@
-package roomescape.reservation.service;
-
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
-import roomescape.reservation.entity.Payment;
-import roomescape.reservation.repository.PaymentRepository;
-
-@Service
-@RequiredArgsConstructor
-public class PaymentService {
-
-    private final PaymentRestClient paymentRestClient;
-    private final PaymentRepository paymentRepository;
-
-    @Transactional
-    public void create(Payment payment) {
-        paymentRepository.save(payment);
-        paymentRestClient.approve(payment);
-    }
-}
