+++ src/test/java/roomescape/helper/TestConfig.java
@@ -3,7 +3,7 @@ package roomescape.helper;
 import org.mockito.Mockito;
 import org.springframework.boot.test.context.TestConfiguration;
 import org.springframework.context.annotation.Bean;
-import roomescape.reservation.service.PaymentService;
+import roomescape.payment.service.PaymentService;
 
 @TestConfiguration
 public class TestConfig {
